backend/internal/scrapers/requirements.txt
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.1
psycopg2-binary>=2.9
python-dotenv>=1.0
selenium>=4.15
brotli>=1.1
